added message sending

+++ internal/domain/models/message.go
@@ -3,9 +3,9 @@ package models
 import "time"
 
 type Message struct {
-	ID        int       `json:"id" db:"id"`
-	ChatID    int       `json:"chat_id" db:"chat_id"`
-	SenderID  int       `json:"sender_id" db:"sender_id"`
+	ID        int64       `json:"id" db:"id"`
+	ChatID    int64       `json:"chat_id" db:"chat_id"`
+	SenderID  int64       `json:"sender_id" db:"sender_id"`
 	Content   string    `json:"text" db:"text"`
 	CreatedAt time.Time `json:"created_at" db:"created_at"`
 	ExpiredAt time.Time `json:"expired_at" db:"expired_at"`
